[Consensus Observer] Make subscription creation asynchronous.
